Add some webhook client tests, rearrange/simplify webhook managter tests, and do some refactoring on the webhook client. Streamify some for loops (using filters for side effects...ew? Also: findFirst short-circuits, so it won't put to all servers), return values from web client methods.
